+++ src/ctl_knob.cpp
@@ -126,7 +126,7 @@ calf_knob_snapshot (GtkWidget *widget, GtkSnapshot *snapshot)
     int width  = gtk_widget_get_width(widget);
     int height = gtk_widget_get_height(widget);
 
-    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, width, height);
+    graphene_rect_t bounds = GRAPHENE_RECT_INIT(0, 0, (float)width, (float)height);
     cairo_t *ctx = gtk_snapshot_append_cairo(snapshot, &bounds);
 
     float r, g, b;
